refactor: 거리 가중치과 시간 가중치 클래스로 분리
